streamlit_env/pyproject.toml
[project]
name = "streamlit-env"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
